+++ pkg/application/doguInstallationUseCase.go
@@ -3,11 +3,11 @@ package application
 import (
 	"context"
 	"fmt"
+
 	cescommons "github.com/cloudogu/ces-commons-lib/dogu"
 	"github.com/cloudogu/k8s-blueprint-operator/v2/pkg/domain"
 	"github.com/cloudogu/k8s-blueprint-operator/v2/pkg/domain/ecosystem"
 	"github.com/cloudogu/k8s-blueprint-operator/v2/pkg/domainservice"
-	"github.com/cloudogu/k8s-blueprint-operator/v2/pkg/util"
 	"golang.org/x/exp/maps"
 	"sigs.k8s.io/controller-runtime/pkg/log"
 )
@@ -16,23 +16,29 @@ type DoguInstallationUseCase struct {
 	blueprintSpecRepo  blueprintSpecRepository
 	doguRepo           doguInstallationRepository
 	waitConfigProvider healthWaitConfigProvider
+	doguConfigRepo     doguConfigRepository
+	globalConfigRepo   globalConfigRepository
 }
 
 func NewDoguInstallationUseCase(
 	blueprintSpecRepo domainservice.BlueprintSpecRepository,
 	doguRepo domainservice.DoguInstallationRepository,
 	waitConfigProvider domainservice.HealthWaitConfigProvider,
+	doguConfigRepo doguConfigRepository,
+	globalConfigRepo globalConfigRepository,
 ) *DoguInstallationUseCase {
 	return &DoguInstallationUseCase{
 		blueprintSpecRepo:  blueprintSpecRepo,
 		doguRepo:           doguRepo,
 		waitConfigProvider: waitConfigProvider,
+		doguConfigRepo:     doguConfigRepo,
+		globalConfigRepo:   globalConfigRepo,
 	}
 }
 
 func (useCase *DoguInstallationUseCase) CheckDoguHealth(ctx context.Context) (ecosystem.DoguHealthResult, error) {
 	logger := log.FromContext(ctx).WithName("DoguInstallationUseCase.CheckDoguHealth")
-	logger.Info("check dogu health...")
+	logger.V(2).Info("check dogu health...")
 	installedDogus, err := useCase.doguRepo.GetAll(ctx)
 	if err != nil {
 		return ecosystem.DoguHealthResult{}, fmt.Errorf("cannot evaluate dogu health states: %w", err)
@@ -41,69 +47,57 @@ func (useCase *DoguInstallationUseCase) CheckDoguHealth(ctx context.Context) (ec
 	return ecosystem.CalculateDoguHealthResult(maps.Values(installedDogus)), nil
 }
 
-func (useCase *DoguInstallationUseCase) WaitForHealthyDogus(ctx context.Context) (ecosystem.DoguHealthResult, error) {
-	logger := log.FromContext(ctx).WithName("DoguInstallationUseCase.WaitForHealthyDogus")
-
-	waitConfig, err := useCase.waitConfigProvider.GetWaitConfig(ctx)
+func (useCase *DoguInstallationUseCase) CheckDogusUpToDate(ctx context.Context) ([]cescommons.SimpleName, error) {
+	logger := log.FromContext(ctx).WithName("DoguInstallationUseCase.CheckDoguHealth")
+	logger.V(2).Info("check if dogus are up to date...")
+	installedDogus, err := useCase.doguRepo.GetAll(ctx)
 	if err != nil {
-		return ecosystem.DoguHealthResult{}, fmt.Errorf("failed to get health check interval: %w", err)
+		return nil, err
 	}
 
-	logger.Info("start waiting for dogu health")
-	healthResult, err := util.RetryUntilSuccessOrCancellation(
-		ctx,
-		waitConfig.Interval,
-		useCase.checkDoguHealthStatesRetryable,
-	)
-	var result ecosystem.DoguHealthResult
-	if healthResult == nil {
-		result = ecosystem.DoguHealthResult{}
-	} else {
-		result = *healthResult
-	}
-
+	globalConfig, err := useCase.globalConfigRepo.Get(ctx)
 	if err != nil {
-		err = fmt.Errorf("stop waiting for dogu health: %w", err)
-		logger.Error(err, "stop waiting for dogu health because of an error or time out")
-	} else {
-		logger.Info("finished waiting for dogu health")
+		return nil, err
+	}
+	globalConfigUpdateTime := globalConfig.LastUpdated
+
+	var dogusNotUpToDate []cescommons.SimpleName
+
+	for doguName, dogu := range installedDogus {
+		versionUpToDate := dogu.IsVersionUpToDate()
+		if !versionUpToDate {
+			dogusNotUpToDate = append(dogusNotUpToDate, doguName)
+			continue
+		}
+
+		doguConfig, err := useCase.doguConfigRepo.Get(ctx, doguName)
+		if err != nil {
+			return nil, err
+		}
+		doguConfigUpdateTime := doguConfig.LastUpdated
+		configUpToDate := dogu.IsConfigUpToDate(globalConfigUpdateTime, doguConfigUpdateTime)
+		if !configUpToDate {
+			dogusNotUpToDate = append(dogusNotUpToDate, doguName)
+			continue
+		}
 	}
 
-	return result, err
-}
-
-func (useCase *DoguInstallationUseCase) checkDoguHealthStatesRetryable(ctx context.Context) (result *ecosystem.DoguHealthResult, err error, shouldRetry bool) {
-	// use named return values to make their meaning clear
-	health, err := useCase.CheckDoguHealth(ctx)
-	if err != nil {
-		// no retry if error while loading dogus
-		return &ecosystem.DoguHealthResult{}, err, false
-	}
-	result = &health
-	shouldRetry = !health.AllHealthy()
-	return
+	return dogusNotUpToDate, nil
 }
 
 // ApplyDoguStates applies the expected dogu state from the Blueprint to the ecosystem.
 // Fail-fast here, so that the possible damage is as small as possible.
-func (useCase *DoguInstallationUseCase) ApplyDoguStates(ctx context.Context, blueprintId string) error {
-	logger := log.FromContext(ctx).WithName("DoguInstallationUseCase.ApplyDoguChanges").
-		WithValues("blueprintId", blueprintId)
-	log.IntoContext(ctx, logger)
-
-	blueprintSpec, err := useCase.blueprintSpecRepo.GetById(ctx, blueprintId)
-	if err != nil {
-		return fmt.Errorf("cannot load blueprint spec %q to install dogus: %w", blueprintId, err)
-	}
-
+func (useCase *DoguInstallationUseCase) ApplyDoguStates(ctx context.Context, blueprint *domain.BlueprintSpec) error {
+	logger := log.FromContext(ctx).WithName("DoguInstallationUseCase.ApplyDoguChanges")
+	logger.V(2).Info("apply dogu states")
 	// DoguDiff contains all installed dogus anyway (but some with action none) so we can load them all at once
 	dogus, err := useCase.doguRepo.GetAll(ctx)
 	if err != nil {
 		return fmt.Errorf("cannot load dogu installations to apply dogu state: %w", err)
 	}
 
-	for _, doguDiff := range blueprintSpec.StateDiff.DoguDiffs {
-		err = useCase.applyDoguState(ctx, doguDiff, dogus[doguDiff.DoguName], blueprintSpec.Config)
+	for _, doguDiff := range blueprint.StateDiff.DoguDiffs {
+		err = useCase.applyDoguState(ctx, doguDiff, dogus[doguDiff.DoguName], blueprint.Config)
 		if err != nil {
 			return fmt.Errorf("an error occurred while applying dogu state to the ecosystem: %w", err)
 		}
@@ -133,6 +127,9 @@ func (useCase *DoguInstallationUseCase) applyDoguState(
 			)
 			return useCase.doguRepo.Create(ctx, newDogu)
 		case domain.ActionUninstall:
+			if doguInstallation == nil {
+				return &domainservice.NotFoundError{Message: fmt.Sprintf("dogu %q not found", doguDiff.DoguName)}
+			}
 			logger.Info("uninstall dogu")
 			return useCase.doguRepo.Delete(ctx, doguInstallation.Name.SimpleName)
 		case domain.ActionUpgrade:
@@ -155,17 +152,9 @@ func (useCase *DoguInstallationUseCase) applyDoguState(
 			logger.Info("update minimum volume size for dogu")
 			doguInstallation.UpdateMinVolumeSize(doguDiff.Expected.MinVolumeSize)
 			continue
-		case domain.ActionUpdateDoguProxyBodySize:
-			logger.Info("update proxy body size for dogu")
-			doguInstallation.UpdateProxyBodySize(doguDiff.Expected.ReverseProxyConfig.MaxBodySize)
-			continue
-		case domain.ActionUpdateDoguProxyRewriteTarget:
-			logger.Info("update proxy rewrite target for dogu")
-			doguInstallation.UpdateProxyRewriteTarget(doguDiff.Expected.ReverseProxyConfig.RewriteTarget)
-			continue
-		case domain.ActionUpdateDoguProxyAdditionalConfig:
-			logger.Info("update proxy additional config for dogu")
-			doguInstallation.UpdateProxyAdditionalConfig(doguDiff.Expected.ReverseProxyConfig.AdditionalConfig)
+		case domain.ActionUpdateDoguReverseProxyConfig:
+			logger.Info("update proxy config for dogu")
+			doguInstallation.UpdateProxyConfig(doguDiff.Expected.ReverseProxyConfig)
 			continue
 		case domain.ActionUpdateAdditionalMounts:
 			logger.Info("update additional mounts")
@@ -179,6 +168,8 @@ func (useCase *DoguInstallationUseCase) applyDoguState(
 	// If this routine did not terminate until this point, it is always an update.
 	if len(doguDiff.NeededActions) > 0 {
 		logger.Info("upgrade dogu")
+		// remove potential pause reconciliation flags here so that the dogu gets updates again
+		doguInstallation.SetReconciliationPaused(false)
 		return useCase.doguRepo.Update(ctx, doguInstallation)
 	}
